requirements.txt
streamlit>=1.35.0
groq>=0.9.0
pinecone-client>=3.0.0
sentence-transformers>=2.7.0
easyocr>=1.7.1
Pillow>=10.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
